more basic database setup
added / upgraded the customer table, inventory table, and address table in the setup file. Created a docker-compose to start up the database container with the setup file. there is some base customer data and inventory data in the setup file. added query to look up customer info in the readme.